crawler/requirements.txt
"""
Crawler requirements — install with:
    pip install -r requirements.txt
"""
# Core HTTP / HTML parsing
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Selenium (for JavaScript-heavy sites)
selenium==4.18.1
webdriver-manager==4.0.1

# PDF text extraction
pdfplumber==0.10.3

# Database / Supabase
psycopg2-binary==2.9.9
supabase==2.3.4

# Email
secure-smtplib           # stdlib wrapper, no separate install needed

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
tenacity==8.2.3          # retry logic

# Logging
colorlog==6.8.2
